+++ src/application/model/distrib/productModel/product/beverage/sugar/all/Coffee.java
@@ -12,8 +12,8 @@ public class Coffee extends CanAddSugar implements NeedPreparation {
     }
 
     @Override
-    public int getPrice() {
-        return 0;
+    public double getPrice() {
+        return 0.0;
     }
 
     @Override
